+++ lib/conjur/api/ldapsync.rb
@@ -22,52 +22,70 @@ require 'conjur/ldap_sync_job'
 
 module Conjur
   class API
-  # @!group LDAP Sync Service
+    # @!group LDAP Sync Service
 
-    # Trigger a LDAP sync with a given profile.
+    # Fetch a Conjur policy that will bring Conjur into sync with the
+    # LDAP server specified by a profile.
     #
-    # @param [Array] args If first element is a Hash, use it as the parameters for the /sync call. Otherwise, assume the caller is using the old calling convention.
-    # @return [Hash] a hash mapping with keys 'ok' and 'result[:actions]'
-    def ldap_sync_now(*args)
+    # @param [String] profile the name of the LDAP server profile 
+    # @param [Hash] options reserved for future use
+    def ldap_sync_policy profile, options = {}
       
-      # Be backward compatible....
-      # This is kind of gross, but changing the interface for this
-      # method didn't seem worth a major version bump.
-      if args[0].instance_of?(Hash)
-        options = args[0]
-      else
-        options = {}
-        options[:config],options[:format],options[:dry_run] = args[0..2]
-        # Old callers expect /sync to be synchronous
-        options[:detach_job] = false
-      end
-
-      opts = credentials.dup.tap{ |h|
-        h[:headers][:accept] = options[:format]
+      headers = credentials.dup.tap {|h|
+        h[:headers][:accept] = 'text/event-stream'
       }
 
-      options[:dry_run] = !!options[:dry_run]
+      options = options.merge(:config_name => profile)
+      url = Conjur.configuration.appliance_url + "/ldap-sync/policy?#{options.to_query}"
 
-      resp = RestClient::Resource.new(Conjur.configuration.appliance_url, opts)['ldap-sync']['sync'].post(options)
+      # Even though we're using SSE to return the policy, fetch the
+      # whole thing at once into a single response. Retrieving it in
+      # chunks doesn't buy us much of anything except more complicated
+      # client code.
+      response = RestClient::Resource.new(url, headers).get
 
-      case options[:format]
-      when 'text/yaml'
-        resp.body
-      when 'application/json'
-        JSON.parse(resp.body)
-      end
+      json = if response.headers[:content_type] == 'text/event-stream'
+               find_policy_event(response) || find_error_events(response)
+             else
+               %Q({"error": {"message": "Unexpected response from server: #{response.body}"}})
+             end
+      JSON.parse(json)
     end
 
-    # Return a list of detached ldap sync jobs
-    def ldap_sync_jobs
-      resource = RestClient::Resource.new(Conjur.configuration.appliance_url, credentials)
-      response = resource['ldap-sync']['jobs'].get
+    # @api private
+    # Get an LDAP sync profile.
 
-      JSON.parse(response.body).map do |job_hash|
-        Conjur::LdapSyncJob.new_from_json(self, job_hash)
-      end
+    # @param [String] profile name 
+    # @param [Hash] options reserved
+    def ldap_sync_show_profile(profile, options = {})
+      resp = RestClient::Resource.new(Conjur.configuration.appliance_url, credentials)['ldap-sync']['config'].get(options)
+      JSON.parse(resp.body)
     end
 
-  # @!endgroup
+    # @api private
+    # Update an LDAP sync profile. 
+    #
+    # ### Note
+    # DO NOT use this method and the UI to update an LDAP sync profile.
+    #
+    # @param [Hash] profile a hash containing the LDAP sync configuration
+    # @param [Hash] options reserved
+    def ldap_sync_update_profile(profile, options = {})
+      options[:json_config] = profile.to_json
+      resp = RestClient::Resource.new(Conjur.configuration.appliance_url, credentials)['ldap-sync']['config'].put(options.to_json, :content_type => 'application/json')
+      JSON.parse(resp.body)
+    end
+
+    # @!endgroup
+    
+    private
+    def find_policy_event(response)
+      response.body.lines.find {|l| l =~ /^data: {"policy":/}.try(:[], 6..-1)
+    end
+
+    def find_error_events(response)
+      response.body.lines.collect {|l| l.match(/^data: ({"error":.*)/).try(:[], 1)}.compact.join("\n")
+    end
+
   end
 end